split skin stuff from r_things into r_skins
what a mess
