+++ src/Sourcefile
@@ -54,6 +54,7 @@ r_patch.c
 r_patchrotation.c
 r_plane.c
 r_segs.c
+r_skins.c
 r_sky.c
 r_splats.c
 r_things.c
