+++ src/r_patch.c
@@ -16,6 +16,7 @@
 #include "doomdef.h"
 #include "r_patch.h"
 #include "r_things.h"
+#include "r_skins.h"
 #include "z_zone.h"
 #include "w_wad.h"
