+++ src/r_things.h
@@ -1,4 +1,4 @@
-// SONIC ROBO BLAST 2
+// SONIC ROBO BLAST 2 KART
 //-----------------------------------------------------------------------------
 // Copyright (C) 1993-1996 by id Software, Inc.
 // Copyright (C) 1998-2000 by DooM Legacy Team.
@@ -18,13 +18,7 @@
 extern "C" {
 #endif
 
-#include "sounds.h"
 #include "r_plane.h"
-#include "r_patch.h"
-
-// "Left" and "Right" character symbols for additional rotation functionality
-#define ROT_L ('L' - '0')
-#define ROT_R ('R' - '0')
 
 // number of sprite lumps for spritewidth,offset,topoffset lookup tables
 // Fab: this is a hack : should allocate the lookup tables per sprite
@@ -66,11 +60,8 @@ fixed_t R_GetSpriteDirectionalLighting(angle_t angle);
 
 fixed_t R_GetShadowZ(mobj_t *thing, pslope_t **shadowslope);
 
-//faB: find sprites in wadfile, replace existing, add new ones
-//     (only sprites from namelist are added or replaced)
-void R_AddSpriteDefs(UINT16 wadnum);
-
 //SoM: 6/5/2000: Light sprites correctly!
+boolean R_AddSingleSpriteDef(const char *sprname, spritedef_t *spritedef, UINT16 wadnum, UINT16 startlump, UINT16 endlump);
 void R_AddSprites(sector_t *sec, INT32 lightlevel);
 void R_AddPrecipitationSprites(void);
 void R_InitSprites(void);
@@ -90,62 +81,12 @@ typedef struct
 
 void R_DrawMasked(maskcount_t* masks, UINT8 nummasks);
 
-boolean R_ThingVisible (mobj_t *thing);
-boolean R_ThingWithinDist (mobj_t *thing, INT32 limit_dist);
+boolean R_ThingVisible(mobj_t *thing);
+boolean R_ThingWithinDist(mobj_t *thing, INT32 limit_dist);
 
-boolean R_ThingIsFullBright (mobj_t *thing);
-boolean R_ThingIsSemiBright (mobj_t *thing);
-boolean R_ThingIsFullDark (mobj_t *thing);
-
-// -----------
-// SKINS STUFF
-// -----------
-#define SKINNAMESIZE 16
-// should be all lowercase!! S_SKIN processing does a strlwr
-#define DEFAULTSKIN "sonic"
-#define DEFAULTSKIN2 "tails" // secondary player
-#define DEFAULTSKIN3 "knuckles" // third player
-#define DEFAULTSKIN4 "eggman" // fourth player
-
-typedef struct
-{
-	char name[SKINNAMESIZE+1]; // INT16 descriptive name of the skin
-	spritedef_t spritedef;
-	spriteinfo_t sprinfo;
-	UINT16 wadnum;
-	char sprite[4]; // Sprite name, if seperated from S_SKIN.
-	skinflags_t flags;
-
-	char realname[SKINNAMESIZE+1]; // Display name for level completion.
-	char hudname[SKINNAMESIZE+1]; // HUD name to display (officially exactly 5 characters long)
-	char facerank[9], facewant[9], facemmap[9]; // Arbitrarily named patch lumps
-
-	// SRB2kart
-	UINT8 kartspeed;
-	UINT8 kartweight;
-	//
-
-	// Definable color translation table
-	UINT8 starttranscolor;
-	UINT8 prefcolor;
-	fixed_t highresscale; // scale of highres, default is 0.5
-
-	// specific sounds per skin
-	sfxenum_t soundsid[NUMSKINSOUNDS]; // sound # in S_sfx table
-
-	boolean localskin;
-	INT32 localnum;
-} skin_t;
-
-extern CV_PossibleValue_t Forceskin_cons_t[];
-
-// had to move those here Zzz...
-INT32 K_GetSkinNum(player_t *player);
-INT32 K_GetMobjSkinNum(const skin_t *skin, boolean local);
-skin_t *K_GetPlayerSkin(player_t *player);
-skin_t *K_GetMobjSkin(const mobj_t *mobj);
-patch_t *K_GetFacePrefix(player_t *player, INT32 skinnum);
-skin_t *K_GetSkinArray(boolean local);
+boolean R_ThingIsFullBright(mobj_t *thing);
+boolean R_ThingIsSemiBright(mobj_t *thing);
+boolean R_ThingIsFullDark(mobj_t *thing);
 
 // -----------
 // NOT SKINS STUFF !
@@ -231,6 +172,8 @@ void R_AllocVisSpriteMemory(void);
 
 UINT8 *R_GetSpriteTranslation(vissprite_t *vis);
 
+void R_InitDrawNodes(void);
+
 // ----------
 // DRAW NODES
 // ----------
@@ -248,71 +191,6 @@ typedef struct drawnode_s
 	struct drawnode_s *next;
 	struct drawnode_s *prev;
 } drawnode_t;
-
-extern INT32 numskins;
-extern INT32 numlocalskins;
-extern INT32 numallskins;
-extern skin_t skins[MAXSKINS];
-extern UINT8 skinstats[9][9][MAXSKINS];
-extern UINT8 skinstatscount[9][9];
-extern UINT8 skinsorted[MAXSKINS];
-
-void sortSkinGrid(void);
-extern skin_t localskins[MAXLOCALSKINS];
-extern skin_t allskins[MAXSKINS+MAXLOCALSKINS];
-
-boolean SetPlayerSkin(INT32 playernum,const char *skinname);
-void SetPlayerSkinByNum(INT32 playernum,INT32 skinnum); // Tails 03-16-2002
-void SetLocalPlayerSkin(INT32 playernum,const char *skinname, consvar_t *cvar);
-INT32 R_SkinAvailable(const char *name);
-INT32 R_AnySkinAvailable(const char *name);
-INT32 R_LocalSkinAvailable(const char *name, boolean local);
-void R_AddSkins(UINT16 wadnum, boolean local);
-
-void R_InitDrawNodes(void);
-
-char *GetPlayerFacePic(INT32 skinnum);
-
-// Functions to go from sprite character ID to frame number
-// for 2.1 compatibility this still uses the old 'A' + frame code
-// The use of symbols tends to be painful for wad editors though
-// So the future version of this tries to avoid using symbols
-// as much as possible while also defining all 64 slots in a sane manner
-// 2.1:    [[ ABCDEFGHIJKLMNOPQRSTUVWXYZ[\]^_`abcdefghijklmnopqrstuvwxyz{|}~   ]]
-// Future: [[ ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789abcdefghijklmnopqrstuvwxyz!@ ]]
-FUNCMATH FUNCINLINE static ATTRINLINE char R_Frame2Char(UINT8 frame)
-{
-#if 1 // 2.1 compat
-	return 'A' + frame;
-#else
-	if (frame < 26) return 'A' + frame;
-	if (frame < 36) return '0' + (frame - 26);
-	if (frame < 62) return 'a' + (frame - 36);
-	if (frame == 62) return '!';
-	if (frame == 63) return '@';
-	return '\xFF';
-#endif
-}
-
-FUNCMATH FUNCINLINE static ATTRINLINE UINT8 R_Char2Frame(char cn)
-{
-#if 1 // 2.1 compat
-	if (cn == '+') return '\\' - 'A'; // PK3 can't use backslash, so use + instead
-	return cn - 'A';
-#else
-	if (cn >= 'A' && cn <= 'Z') return cn - 'A';
-	if (cn >= '0' && cn <= '9') return (cn - '0') + 26;
-	if (cn >= 'a' && cn <= 'z') return (cn - 'a') + 36;
-	if (cn == '!') return 62;
-	if (cn == '@') return 63;
-	return 255;
-#endif
-}
-
-FUNCMATH FUNCINLINE static ATTRINLINE boolean R_ValidSpriteAngle(UINT8 rotation)
-{
-	return ((rotation <= 8) || (rotation == ROT_L) || (rotation == ROT_R));
-}
 
 #ifdef __cplusplus
 } // extern "C"
